+++ src/core/context/client-file.ts
@@ -8,11 +8,7 @@ export const CLIENT_FILE_NAME = 'client.json'
 export interface ClientInfo {
   clientId: Uint8Array
   /**
-   * LoginId of the account which is under duress and is being impersonated
-   * by the duress account. It should be the loginId which the duress account
-   * is nested under..
-   * This is only set if duress mode is activated via pin-login with the
-   * duress account's pin.
+   * This is a boolean flag that puts the device into duress mode.
    */
   duressEnabled: boolean
 }